+++ src/plugins/v-tooltip/index.js
@@ -1,4 +1,4 @@
 import Vue from "vue";
 import VTooltip from "v-tooltip";
 
-Vue.use(VTooltip, { defaultContainer: "body" });
+Vue.use(VTooltip, { defaultContainer: "body", defaultBoundariesElement: "body" });